+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Now it performs a non-blocking connection. Previous behaviour with `connect_sync
 - Added `NetworkController::is_ready()`
 - Added `borrow()` method for `StoredNetEvent` to transform in into `NetEvent`.
 - Added `is_local()` and `is_remote()` helpers to `ResourceId`.
+- Added `SendStatus::ResourceNotAvailable`
 - Modified `SendStatus::MaxPacketSizeExceeded`, now it not contains lengths info.
 - Renamed `udp::MAX_COMPATIBLE_PAYLOAD_LEN` to `udp::MAX_LOCAL_PAYLOAD_LEN`, value updated
 with `udp::MAX_PAYLOAD_LEN`.